v4.3.0
added Esp32CamAccess_Servo.ino to connect ESP32-CAM with servo because it was possible but ESP32-CAM was unable to run the code due to the performance/memory issue. So basically the specs are not compatible to run much of API and works. Such as HTTP Web Server + Camera Stream/Capture + API (Stream, Capture, Control

At this point, I decided to keep this code for my checkpoint so I can configure the best optimization setting for the device itself later